Repository implementation

+++ LvivCompany.Bookstore/LvivCompany.Bookstore.Entities/Models/BookAuthor.cs
@@ -8,11 +8,11 @@ namespace LvivCompany.Bookstore.Entities
     public class BookAuthor
     {
        
-        public int BookId { get; set; }
+        public long BookId { get; set; }
         public Book Book { get; set; }
 
 
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public Author Author { get; set; }
     }
 }
